Updated instructions and cluster code
Added instructions to initial cleaning script and more added to clustering code.  Reading in the csv's from tool and making appropriate edits to merge  tables.  The tool won't allow different column identifiers, so to be able to do those anovas and have the factors, the .txt files needed to be uploaded for each combination of factors.